refactoring notFound folder structure
