+++ src/js/modules/auth/index.js
@@ -2,14 +2,12 @@
  * Authentication Feature
  */
 import { combineReducers } from 'redux';
-import components from './components';
-import containers from './containers';
+import * as ui from './ui';
 import { actionTypes, actions } from './actions';
 import actionCreators from './actionsCreators';
 import * as reducers from './reducers';
 import routes from './routes';
 
-const rootReducer = combineReducers(reducers);
+const reducer = combineReducers(reducers);
 
-export default rootReducer;
-export { components, containers, actionTypes, actions, routes, actionCreators };
+export { ui, actionTypes, actions, routes, actionCreators, reducer };
