+++ src/js/modules/notFound/routes.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import path from 'path';
 import { Route } from 'react-router-dom';
-import NotFoundView from './components';
+import NotFoundView from './ui/components/NotFoundView';
 
 
 export default function notFoundRoutes(basePath) {
