+++ src/js/modules/notFound/ui/index.js
@@ -1,0 +1,3 @@
+import NotFoundView from './components/NotFoundView';
+
+export const components = { NotFoundView };
